+++ pcdet/ops/pointnet2/pointnet2_stack/pointnet2_utils.py
@@ -80,7 +80,7 @@ class GroupingOperation(Function):
         B = idx_batch_cnt.shape[0]
         output = torch.cuda.FloatTensor(M, C, nsample)
 
-        pointnet2.group_points_wrapper(B, M, C, nsample, features, features_batch_cnt, idx, idx_batch_cnt, output)
+        pointnet2.group_points_wrapper(B, M, C, nsample, features.float(), features_batch_cnt, idx, idx_batch_cnt, output)
 
         ctx.for_backwards = (B, N, idx, features_batch_cnt, idx_batch_cnt)
         return output
